crowd-edge-cloud/functions/analyst/requirements.txt
flask==3.0.2
flask-cors==4.0.0
gunicorn==21.2.0
google-cloud-firestore>=2.11.0
pandas
scikit-learn
numpy